Don't choke when downloading a URL containing non-UTF-8 characters (OLPC#8857)
A URL can contain arbitrary characters; there's no guarantee they are encoded
using UTF-8. Using the encoding of the page that contains the URL is as good a
guess as any.

Instead of throwing an exception and failing to download the file, we will now
record a "garbled" file name in the Journal if there's an encoding mismatch.

Tested-By: Simon Schampijer <simon@schampijer.de>
Reviewed-By: Sascha Silbe <sascha-pgp@silbe.org>
Acked-By: Lucian Branescu <lucian.branescu@gmail.com>