Remove unfinished downloads
When Browse starts removes all files in the instance directory that where not modified in a long time(1 day).
Since there is no way to pause a download and continue later, this files are failed downloads that should be eliminated.
this is the ''cache'' case of an already fixed issue, by Jose Prous joseprous@gmail.com