Browse: tabs usability improved
The Add Tab button has been relocated next to the tab labels, allowing
more space for the URL entry.

Tabs are always shown. There is at least one tab. In that case, it
cannot be closed. We prevent the closing by hiding the 'X' button.
Also, the close image was sugarized.

There is now a fixed width for tabs. The label text does ellipsize.
The width depends on the amount of tabs. There is a maximum size that
is used when there is extra space. There is a minimum size to prevent
hiding the information.

When a new tab is opened, it now shows an empty page, not the default
page. In the future, we will add a hint in this empty page, similar
to what we have for an empty Journal.

Added the option to "follow link in new tab" in the link's palette.
Added an icon for this item, and also the icon for "follow link" is
also updated.

When a new tab is opened via the '+' button, the focus goes to the URL
entry in the toolbar.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>