Use gtk-notebook-action-widget for the 'add-tab' button
Due to API changes in the GTK Notebook [1] we can add now an
action widget to it (e.g. a button to add a new tab). So far our
'add-new-tab'-button in Browse was another special tab. This
does simplify the Browse code a lot by using the action widget
instead. Yay!

There will be a visual difference with this approach, the
Button will always be at the far right, but this is
acceptable - Safari uses the same style, with a button at the
far right for adding a new tab (once a tab is open).

[1] http://developer.gnome.org/gtk3/3.0/GtkNotebook.html#gtk-notebook-set-action-widget

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>