Append the current Sugar (Sucrose) version to the user agent
This identifier is used in ASLO to determine which activities
can be downloaded (ASLO is parsing the user agent of the Browser
to provide an appropriate activity version, if it fails to get
the Sugar version from the agent string, it uses the last stable version,
which is 0.94 at the moment).

One thing that is not perfectly handled by this patch is that we
need to modify the user agent for every WebView (see [1]).

Furthermore we need to modify the Sugar version in Browse accordingly
with each Sugar release as the Sugar version is only availble in the
shell and not in the toolkit.

[1] https://lists.webkit.org/pipermail/webkit-gtk/2012-January/000893.html

Signed-off-by: Simon Schampijer <simon@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>