Show loading message in the overlay
While Browse is loading, the overlay at the bottom-left of the
screen -which is currently used to show the URL of a link when
the mouse is over it- will show "Loading...", and later will
append the requested URI to that message.  This message replaces
the one showing the link URL until the page load is done.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>