Make the LinkInfo that is displayed at the bottom left themable
Add a gtype_name to make it easily accessible from the theme. We have
to do the same for the overlay as well in order to be able to
set the 'background-color' of the label. Labels don't create their own
window to render themselves into. Instead, they draw themselves
directly onto their parents window, so we have to set the background-color
of the parent in order to set the labels' background color [1].

The patch does adjust the gtype_names to be consistent.

[1] http://developer.gnome.org/gtk-faq/stable/x810.html

Signed-off-by: Simon Schampijer <simon@schampijer.de>
Acked-by: Manuel Quiñones <manuq@laptop.org>