Basic PDF reading
When a request is made with a MIME type 'application/pdf', a new tab
is opened next to the current one to show the document.  If the
document is remote, a download starts, and the progress is shown in
the toolbar's URL entry.  It can be cancelled with the X button, the
same for cancelling the loading.

Basic navigation is provided to allow a quick read of the PDF.  The
user can also save the document to the Journal to store it, and later
it may be opened in Read for featureful usage.  The controls are in a
floating toolbar, at the bottom of the document.

The PDF tabs are restored between sessions as expected.

If the user clicks on a hyperlink in a PDF, a new tab is opened next
to the current to browse it.

This uses Evince, the same backend to display the PDF document as
Read.  And WebKit to download the document, as per other downloads in
Browse.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>