Update progress bar and stop/reload buttons when switching tabs
Calling the _set_status and _set_progress methods in the web toolbar
directly when a new tab is activated.

Also the callback for load-status is disconnected now in the previous
tab, like the other callbacks.

Removed the update of the navigation buttons when the loading status
changes.  This is only needed when the adress changes or when the tabs
are switched.

Removed the title set to None when loading status chenges, this was
giving the following Gtk error:

Gtk-CRITICAL **: gtk_entry_set_text: assertion `text != NULL' failed

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>