Store empty strings instead of None in the places sqlite database.
The SQL model sets "text" as the fields type for uri and title.
However, sqlite uses a dynamic type sistem [1] and our Place class was
initializing those attributes with None.  As a consequence, when the
user types something in the toolbar entry, and a search is made to
show a drop-down list for autocomplete, the drop-down list is being
filled with None, but is defined as Gtk.ListStore(str, str).  This is
provoking a GTK+ error,that crashes the activity in the XO, as bug

In order to keep backwards compatibility, if the database returns None
for does fields, they are translated to empty strings in
_place_from_row() method of place.SqliteStore .

[1] http://sqlite.org/datatype3.html

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Signed-off-by: Simon Schampijer <simon@laptop.org>