Pass URI instead of file path when the activity starts from a file
If the MIME type of the file that the activity reads is not
"text/plain" or "text/uri-list", it will load the file as-is in the
default tab.  But WebKit needs an URI, and we are passing the file
path.  Appending 'file://' to the file path solves the issue.

This fixes second part of #3300 .

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>