Set default-font-size and monospace-font-size in WebKit based on GtkSettings
The font size that GtkSettings provides is multiplied to reach the
default values in WebKitWebSettings in a common 96dpi display.  Those
are 12 for default-font-size and 10 for monospace-font-size, so the
monospace font is calculated as 2 units less from the default Gtk+
font.

This is similar to Epiphany approach, when the "Use system fonts"
is checked in the preferences dialog.

This workarounds #3428 "Browse should set DPI for XO screen", tested
with screenshots that the resulting font sizes are similar to the
previous Browse based on Gecko in the XO.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Daniel Drake <dsd@laptop.org>