Store web session cookies in SQlite database, for SL #3456
There was code in Browse to create a SQlite database under data
directory in the activity profile.  This was for creating a cookie to
authenticate the laptop with a schoolserver.

Now we use the same database via the Soup library to store all cookies
so they are persistent between sessions.  A Soup cookie jar is being
attached to the WebKit session, cookies are stored in a Mozilla
compatible SQLite format.

The schoolserver function has been changed to use the Soup API.

The code is based in Epiphany, embed/ephy-embed-single.c .

------

v1->v2: let Soup take care of database creation, use Soup API in
schoolserver cookie seed function

v2->v3: use proper constructors for PyGI instead of new()

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Daniel Drake <dsd@laptop.org>