Handle new windows request and instead open links in new tabs #3507
Browse being a Sugar activity only has one full screen window.  This
patchs adds a callback to the new window policy decision requested
signal and handles new links in new tabs.

If in need, we can also take the decision to not open the link, using:

reason = navigation_action.get_reason()

I have not come to a reason to do that yet.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>