Display 'Loading...' message in the tab and url titles while the page is loading #3550
First of all, we don't need to open "about:blank" when we create a new
tab, this was making the history to save that "page" and allowing to
go back to it.

That is related with this PATCH because we are checking the browser
"load-status" property to know what title put in the tab and url
label. Now, we check for the "load-status" property and set the title
according to the load progress: if WebKit is loading the page we put
"Loading..."  or the page's title otherwise with the exception of
"about:blank" page that has no title, so we put "Untitled".

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Acked-by: Manuel Quiñones <manuq@laptop.org>
Acked-by: Simon Schampijer <simon@laptop.org>