Unset "loading..." message when the loading fails SL #3620
Browse has to restore the previous title in the tab and in the URL
entry when the loading finishes with WEBKIT_LOAD_FAILED status.

This fixes bug #3620 in Sugarlabs trac.  The WEBKIT_LOAD_FAILED
status is reached in the current webview when the loading is
manually handled in the mime-type-policy-decision-requested callback.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Tested-by: Manuel Kaufmann <humitos@gmail.com>
Acked-by: Simon Schampijer <simon@laptop.org>