Restore journal file picker (#3411)
The run-file-chooser signal can be used to intercept a request
for a file upload and display a custom file chooser.

We want to do this to allow the user to upload files from the journal.
By default, the standard GTK+ file chooser was being shown.

run-file-chooser is present in WebKit2 and has been submitted for
final review for inclusion in WebKit1.

Signed-off-by: Daniel Drake <dsd@laptop.org>
Acked-by: Manuel Quiñones <manuq@laptop.org>