Restore the Bookmarks tray handling using a toggle button - SL #3868
Instead of changing the tooltip and icon of the tool button based on
the visibility of the tray, with the map and unmap callbacks that this
patch removes, do the inverse: when a bookmark is made, the toggle
button is activated, and it shows the tray in the toggled callback.
Also when the only pending bookmark is removed, the toggle button is
deactivated, and it hides the tray in the toggled callback.

The switch to a toggle button allows us to do the tray and button
interaction in a more standard way, and the bugs that the
unconventional interaction raised are now fixed.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>
Signed-off-by: Manuel Kaufmann <humitos@gmail.com>