Handle LoadInterruptedByPolicyChange SL #3972
This error is raised when the mimetype policy suffers a change. When
this happens __load_error_cb is called and we were showing the error
page in any case.

With this patch that case is properly handled in __load_error_cb and
the error page is not shown on the current tab. In fact, we keep the
same page where the user clicked the link in the tab.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Acked-by: Manuel Quiñones <manuq@laptop.org>