New layout of tabs SL #3560
When Browse starts it shows just one tab sized half of the whole
canvas. Every time a new tab is added they are resized and expanded
over the full canvas width to make them fit. Once the 9th[1] tab is
reached, all of them keep the same size and two buttons are added
to be able to scroll over them.

[1] this decision was taken here 959f86e0406914267047dc10302f98a059b6ba21

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Acked-by: Manuel Quiñones <manuq@laptop.org>