Title of media files SL #3934
When a plugin handles the request, for example to play an .ogg file,
WebKit sets the title as 'None' and '__load_status_changed_cb' is not
called when status is 'WebKit.LoadStatus.FINISHED', so the title is
not set as 'Untitled'.

This patch checks if the title is 'None' (because of the case
mentioned above) and sets it as the 'os.path.basename' of the
'WebKit.WebView.props.uri' (the name of the filename).

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Acked-by: Manuel Quiñones <manuq@laptop.org>