Grab the WebKit.WebView focus on tab closing #4187
When the current tab is closed the next tab is shown (left one if the
current tab was the last, othercase right one). After this the
WebKit.WebView of the tab displayed grabs the focus to show the OSK if
this tab was showing it.

So, switching between tabs and closing them respect the OSK status
(shown / hidden).

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Acked-by: Manuel Quiñones <manuq@laptop.org>