Handle 'create-web-view' signal and open a new tab SL #4236
There are some cases that WebKit.WebView needs to create a new
WebKit.WebView (e.g. to open a pop up window) and it emits this
signal. Handling this signal we are able to insert a new tab (instead
of creating a new window) and attach it next to the current one with
the desired url to be load.

The signal 'new-window-policy-decision-requested' was removed because
it is not more needed because 'create-web-view' will work for all kind
of popups.

Signed-off-by: Manuel Kaufmann <humitos@gmail.com>
Acked-by: Manuel Quiñones <manuq@laptop.org>