Partial fix to style the Embedded media player - SL #3933 #4465
The only way to do this in webkitgtk1 is to inject CSS using
WebKit.WebView.execute_script:

http://webkitgtk.org/reference/webkitgtk/stable/webkitgtk-webkitwebview.html#webkit-web-view-execute-script

The new CSS rules do:

- set page background white
- set player background dark grey, same as Sugar toolbar
- set player text white
- set buttons size to the same size as Sugar toolbar buttons

This commit doesn't change the size of the icons inside the buttons,
which is too small.  Ongoing investigation in the #4465 ticket, it
looks like it should be an upstream fix.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>