Palette: join the different pallettes in one palette with sections - SL #4262
Previously the code used three different palettes.  If the cursor was
over a link, it displayed the link palette with options like "Follow
link".  If that wasn't the case, it checked for image, and displayed
options to act on it like "Keep image".  Otherwise, if the cursor was
over a user selected text, it displayed the palette to "Copy text".

But as the bug #4262 state, those cases are not exclusive.  For
example, the user can pop-up the palette over an image which is also a
link, and also user selected.  In that case, all the options should be
displayed.  This commit does so.

Signed-off-by: Manuel Quiñones <manuq@laptop.org>