Port load/read file to webkit2
This requires using the new session state APIs.  We can maintain
backwards compatibility in the form of writing the old format
history and in the form of reading the current url from an old
format history.  However, the code prefers to load from the new
session state format rather than the old format (because it will
put both into a file for backwards compat).

This requires 2.11.3+ (for session state apis).  It seems to work
on 2.11.5, but it is hard to test due to the lack of general stability
in that release.