Fix ctrl+c and ctrl+v
Problem: ctrl+c and ctrl+v do not work in the URL entry, but emit an
alert sound.

Analysis: the key-press-event handlers for the activity, the toolbar,
and the entry widget are coded to pass unhandled events to the Gtk.Entry
widget, and this happens properly when sugar3.IconEntry is used with the
example, but not when used with the activity.  Cause unknown.

Solution: explicitly handle the ctrl+c and ctrl+v events.

References:
https://bugs.sugarlabs.org/ticket/4979
https://bugs.sugarlabs.org/ticket/4642

[explicit handling of events]
Written-by: GAURAV MISHRA <gauravm043@gmail.com>

[test, review, and switch to Gdk.KEY_* constants]
Signed-off-by: James Cameron <quozl@laptop.org>