Open with an activity if possible #4981
Based on 7315fe7 and a56d993 which were lost in port to WebKit2.

Tries to obtain the bundle that will be invoked for the downloaded
saved journal object.

Alert will say either "Open with %s", where %s is the name of the
activity, or to "Show in Journal".

Fixes https://bugs.sugarlabs.org/ticket/4981