Fix image download
Due to a bug in the preview generation, image download is currently
broken in master (the python3) version. Issue is using io.StringIO,
where io.BytesIO should be used. A preview is being generated, but size
of image in preview is small. That's due to another issue though,
haven't debugged that yet.

below patch fixes the download issue for me. Tested with python3 (2 not
revelant, as master is for python3 only, right?) and download of images
works again.

best regards,
Bas Hulsken

Fixes https://github.com/sugarlabs/browse-activity/issues/107

Signed-off-by: James Cameron <quozl@laptop.org>