updated 'btn' class to have cursor: pointer so that it feels interactable
The `btn` class is used on an input tag which is supposed to act as a button. 
buttons generally have some hover effect or some sort of indication that they're clickable. Now, it is quite obvious that the `Google Search` text means that it's a button but just so that it feels a little more polished I added the `cursor:pointer` just like there is on the sugarlabs logo.